requirements.txt
discord.py==2.6.4
python-dotenv==1.0.0
Flask==3.0.0
requests==2.31.0